requirements.txt
discord.py
httpx
openai>=1.35.0
pyyaml
